docs/validation/requirements.txt
pyspellchecker==0.7.2
pyyaml==6.0.1
regex==2023.12.25 